add font, refactor components folder structure
