+++ components/Editor/index.ts
