+++ components/LayoutMain/index.ts
@@ -1,0 +1,2 @@
+
+export * from "./LayoutMain"
